+++ test/videouplinkbandwidthpolicy/NScaleVideoUplinkBandwidthPolicy.test.ts
@@ -488,6 +488,53 @@ describe('NScaleVideoUplinkBandwidthPolicy', () => {
         );
       }
     });
+
+    it('Scale to 360 if browser behavior returns disable 480p', () => {
+      class TestBrowserBehavior extends DefaultBrowserBehavior {
+        disable480pResolutionScaleDown(): boolean {
+          return true;
+        }
+      }
+      policy = new NScaleVideoUplinkBandwidthPolicy(
+        selfAttendeeId,
+        true,
+        undefined,
+        new TestBrowserBehavior()
+      );
+      policy.setIdealMaxBandwidthKbps(600);
+      policy.setTransceiverController(transceiverController);
+      transceiverController.setVideoInput(new MediaStreamTrack());
+      const numParticipants = 5;
+      const expectedParams = new DefaultVideoCaptureAndEncodeParameter(
+        320,
+        192,
+        15,
+        320,
+        false,
+        1.5
+      );
+      const sources: SdkStreamDescriptor[] = [];
+      for (let i = 0; i < numParticipants; i++) {
+        const attendee = i === 0 ? selfAttendeeId : `attendee-${i}`;
+        sources.push(
+          new SdkStreamDescriptor({
+            streamId: i,
+            groupId: i,
+            maxBitrateKbps: 100,
+            attendeeId: attendee,
+            mediaType: SdkStreamMediaType.VIDEO,
+          })
+        );
+      }
+      const index = new DefaultVideoStreamIndex(logger);
+      index.integrateIndexFrame(new SdkIndexFrame({ sources: sources }));
+      policy.updateIndex(index);
+      const actualParams = policy.chooseCaptureAndEncodeParameters();
+      assert(
+        actualParams.equal(expectedParams),
+        `Expected: ${JSON.stringify(expectedParams)} actual: ${JSON.stringify(actualParams)}`
+      );
+    });
   });
 
   describe('wantsResubscribe', () => {